修复 double 转 int 警告

+++ NatureEnhance/DataStruct.cpp
@@ -5,8 +5,8 @@ expReal ne_list_create(GMString info)
 {
 	int id = gm::ds_list_create();
 
-	gm::ds_map_add(PropertyMap, id, ds_type_list);
-	gm::ds_map_add(NameMap, id, info);
+	gm::ds_map_add((int)PropertyMap, id, ds_type_list);
+	gm::ds_map_add((int)NameMap, id, info);
 
 	return id;
 }
@@ -22,8 +22,8 @@ expReal ne_list_destroy(GMReal id)
 
 		gm::ds_list_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i);
-		gm::ds_map_delete(NameMap, i);
+		gm::ds_map_delete((int)PropertyMap, i);
+		gm::ds_map_delete((int)NameMap, i);
 
 		finish;
 	}
@@ -34,8 +34,8 @@ expReal ne_map_create(GMString info)
 {
 	int id = gm::ds_map_create();
 
-	gm::ds_map_add(PropertyMap, id + 100000000, ds_type_map);
-	gm::ds_map_add(NameMap, id + 100000000, info);
+	gm::ds_map_add((int)PropertyMap, id + 100000000, ds_type_map);
+	gm::ds_map_add((int)NameMap, id + 100000000, info);
 
 	return id;
 }
@@ -51,8 +51,8 @@ expReal ne_map_destroy(GMReal id)
 
 		gm::ds_map_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i + 100000000);
-		gm::ds_map_delete(NameMap, i + 100000000);
+		gm::ds_map_delete((int)PropertyMap, i + 100000000);
+		gm::ds_map_delete((int)NameMap, i + 100000000);
 
 		finish;
 	}
@@ -63,8 +63,8 @@ expReal ne_stack_create(GMString info)
 {
 	int id = gm::ds_stack_create();
 
-	gm::ds_map_add(PropertyMap, id + 200000000, ds_type_stack);
-	gm::ds_map_add(NameMap, id + 200000000, info);
+	gm::ds_map_add((int)PropertyMap, id + 200000000, ds_type_stack);
+	gm::ds_map_add((int)NameMap, id + 200000000, info);
 
 	return id;
 }
@@ -80,8 +80,8 @@ expReal ne_stack_destroy(GMReal id)
 
 		gm::ds_stack_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i + 200000000);
-		gm::ds_map_delete(NameMap, i + 200000000);
+		gm::ds_map_delete((int)PropertyMap, i + 200000000);
+		gm::ds_map_delete((int)NameMap, i + 200000000);
 
 		finish;
 	}
@@ -92,8 +92,8 @@ expReal ne_queue_create(GMString info)
 {
 	int id = gm::ds_queue_create();
 
-	gm::ds_map_add(PropertyMap, id + 300000000, ds_type_queue);
-	gm::ds_map_add(NameMap, id + 300000000, info);
+	gm::ds_map_add((int)PropertyMap, id + 300000000, ds_type_queue);
+	gm::ds_map_add((int)NameMap, id + 300000000, info);
 
 	return id;
 }
@@ -109,8 +109,8 @@ expReal ne_queue_destroy(GMReal id)
 
 		gm::ds_queue_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i + 300000000);
-		gm::ds_map_delete(NameMap, i + 300000000);
+		gm::ds_map_delete((int)PropertyMap, i + 300000000);
+		gm::ds_map_delete((int)NameMap, i + 300000000);
 
 		finish;
 	}
@@ -121,8 +121,8 @@ expReal ne_grid_create(GMReal w, GMReal h, GMString info)
 {
 	int id = gm::ds_grid_create(static_cast<int>(w), static_cast<int>(h));
 
-	gm::ds_map_add(PropertyMap, id + 400000000, ds_type_grid);
-	gm::ds_map_add(NameMap, id + 400000000, info);
+	gm::ds_map_add((int)PropertyMap, id + 400000000, ds_type_grid);
+	gm::ds_map_add((int)NameMap, id + 400000000, info);
 
 	return id;
 }
@@ -138,8 +138,8 @@ expReal ne_grid_destroy(GMReal id)
 
 		gm::ds_grid_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i + 400000000);
-		gm::ds_map_delete(NameMap, i + 400000000);
+		gm::ds_map_delete((int)PropertyMap, i + 400000000);
+		gm::ds_map_delete((int)NameMap, i + 400000000);
 
 		finish;
 	}
@@ -150,8 +150,8 @@ expReal ne_priority_create(GMString info)
 {
 	int id = gm::ds_priority_create();
 
-	gm::ds_map_add(PropertyMap, id + 500000000, ds_type_priority);
-	gm::ds_map_add(NameMap, id + 500000000, info);
+	gm::ds_map_add((int)PropertyMap, id + 500000000, ds_type_priority);
+	gm::ds_map_add((int)NameMap, id + 500000000, info);
 
 	return id;
 }
@@ -167,8 +167,8 @@ expReal ne_priority_destroy(GMReal id)
 
 		gm::ds_priority_destroy(i);
 
-		gm::ds_map_delete(PropertyMap, i + 500000000);
-		gm::ds_map_delete(NameMap, i + 500000000);
+		gm::ds_map_delete((int)PropertyMap, i + 500000000);
+		gm::ds_map_delete((int)NameMap, i + 500000000);
 
 		finish;
 	}
@@ -180,18 +180,18 @@ expReal ne_list_read_buffer(GMReal list, GMReal buffer, GMReal types)
 	try
 	{
 		int listSize = static_cast<int>(gm::buffer_read_uint32(buffer));
-		int typeSize = gm::ds_list_size(types);
+		int typeSize = gm::ds_list_size((int)types);
 
 		if (listSize % typeSize != 0)
 			throw L"被 buffer 记载的 list 的大小不能被 types 的大小整除。";
 
-		gm::ds_list_clear(list);
+		gm::ds_list_clear((int)list);
 		for (int i = 0; i < listSize; ++i)
 		{
-			GMReal type = std::get<GMReal>(gm::buffer_read(buffer, 
-				gm::ds_list_find_value(types, i % typeSize)));
+			GMReal type = std::get<GMReal>(gm::buffer_read((int)buffer,
+				(int)gm::ds_list_find_value((int)types, i % typeSize)));
 
-			gm::ds_list_add(list, type);
+			gm::ds_list_add((int)list, type);
 		}
 
 		ne_list_destroy(types);
@@ -204,8 +204,8 @@ expReal ne_list_write_buffer(GMReal list, GMReal buffer, GMReal types)
 {
 	try
 	{
-		int listSize = gm::ds_list_size(list);
-		int typeSize = gm::ds_list_size(types);
+		int listSize = gm::ds_list_size((int)list);
+		int typeSize = gm::ds_list_size((int)types);
 
 		if (listSize % typeSize != 0)
 			throw L"传入的 list 的大小不能被 types 的大小整除。";
@@ -213,9 +213,9 @@ expReal ne_list_write_buffer(GMReal list, GMReal buffer, GMReal types)
 		gm::buffer_write_uint32(buffer, listSize);
 		for (int i = 0; i < listSize; ++i)
 		{
-			GMReal type = gm::ds_list_find_value(types, i % typeSize);
-			gm::CGMVariable value = gm::ds_list_find_value(list, i);
-			gm::buffer_write(buffer, type, value);
+			GMReal type = gm::ds_list_find_value((int)types, i % typeSize);
+			gm::CGMVariable value = gm::ds_list_find_value((int)list, i);
+			gm::buffer_write((int)buffer, (int)type, value);
 		}
 
 		ne_list_destroy(types);
@@ -239,10 +239,10 @@ expReal ds_exists(GMReal id, GMReal type)
 	else if (type == ds_type_priority)
 		i += 500000000;
 
-	if (!gm::ds_map_exists(PropertyMap, i))
+	if (!gm::ds_map_exists((int)PropertyMap, i))
 		return 0;
 	
-	return gm::ds_map_find_value(PropertyMap, i) == type ? 1 : 0;
+	return gm::ds_map_find_value((int)PropertyMap, i) == type ? 1 : 0;
 }
 
 template <typename... Args>
